taking a new approach, refactoring templates to use lisp-like syntax
The new template syntax resembles both the syntax of lisp and Chomsky's phrase marker syntax. The former resemblance is intentional and the latter resemblance is unintentional. The advantaged offered by this approach is that the syntax is easy to implement and process yet allows the templates represented by the nested lists to be trivially modified. This allows us to cleanup and restructure the inflection of utterances in such a way that some operations can be shared between both native and foreign languages. More importantly though, it accelerates our ability to implement the remaining functionality that's needed such as for adpositions, pronoun possessives, adjective agreement, etc., since templates for declension can now be repurposed to these needs by applying a series of operations to the nested lists that represent them. Conjugation works under this new interpretation, however declensions for common nouns and pronouns still need to be reimplemented.