renaming: LanguageScript→Writing, & introducing ∅ symbol to tables
"Writing" and its usage are arguably more idiomatic, and are definitely closer to the interfaces that have come out of past design.
The empty set symbol "∅" is useful to distinguish situations where a string should not render due to a necessary conjugation/declension being non-existent/not-applicable, vs situations where a string should render but should literally represent the conjugation/declension with an empty string. The empty set symbol removes the need for the `omit_empty` parameter, which complicated the logic and could only determine behavior on a table-by-table basis