overhaul to remove predicate library
The data structures of `dictstores` can be used to supplant the predicate library, and it does so in a much simpler way that's easier to reason with and debug, while also reusing code that was developed for other functionality, and obviating code that was used to distinguish declension decks from conjugation decks. win! (just as long as it can reach parity with old code, emoji demonstration is still disabled)