definiteness is a tagaxis, nps are ordered, closures respect ordering
Articles were broken for dummy nouns that had to be indefinite. To fix articles, definiteness would need to be expressed as a tagaxis, which would shore up tech debt needed to support languages like arabic. English translations would express definiteness by adding articles to the tree. However, since ordering for noun phrases was not yet implemented, this caused the inserted articles to be misordered. Fixing this required implementing ordering for noun phrases, which would be necessary anyway to support languages with different adjective orderings, such as Spanish. However, this would not be possible for as long as start and end delimiters (such as closures, parentheses, and brackets) were treated as nodes within syntax trees, since the contents of such delimiters would not be easily accessible during the ordering of noun phrases. To fix this, delimited text regions would need to be represented using tags, which had been planned for a while  anyway, since it would be necessary to avoid [awkward] [bracketed] [statements] [like] [this] in more complicated text for demonstrating things such as the middle voice